feat: Money 변경 적용

+++ src/main/java/ex/sample/domain/sample/dto/request/CreateSampleReq.java
@@ -2,7 +2,7 @@ package ex.sample.domain.sample.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 public record CreateSampleReq(
 
@@ -10,7 +10,7 @@ public record CreateSampleReq(
     String name,
 
     @NotNull(message = "금액은 필수값입니다.")
-    BigDecimal money
+    BigInteger money
 ) {
 
 }
